:recycle: Refactor frontend state, integrate typesafe-vuex
accessors into state files
