--- {{cookiecutter.project_slug}}/frontend/src/store/admin/accessors/commit.ts
@@ -1,9 +1,0 @@
-import { getStoreAccessors } from 'typesafe-vuex';
-import { State } from '@/store/state';
-import { mutations } from '../mutations';
-import { AdminState } from '../state';
-
-const {commit} = getStoreAccessors<AdminState, State>('');
-
-export const commitSetUser = commit(mutations.setUser);
-export const commitSetUsers = commit(mutations.setUsers);
