--- {{cookiecutter.project_slug}}/frontend/src/store/admin/accessors/dispatch.ts
@@ -1,10 +1,0 @@
-import { getStoreAccessors } from 'typesafe-vuex';
-import { AdminState } from '../state';
-import { State } from '@/store/state';
-import { actions } from '../actions';
-
-const {dispatch} = getStoreAccessors<AdminState, State>('');
-
-export const dispatchCreateUser = dispatch(actions.actionCreateUser);
-export const dispatchGetUsers = dispatch(actions.actionGetUsers);
-export const dispatchUpdateUser = dispatch(actions.actionUpdateUser);
