--- {{cookiecutter.project_slug}}/frontend/src/store/admin/accessors/index.ts
@@ -1,3 +1,0 @@
-export * from './commit';
-export * from './dispatch';
-export * from './read';
