--- {{cookiecutter.project_slug}}/frontend/src/store/admin/accessors/read.ts
@@ -1,9 +1,0 @@
-import { getStoreAccessors } from 'typesafe-vuex';
-import { AdminState } from '../state';
-import { State } from '@/store/state';
-import { getters } from '../getters';
-
-const { read } = getStoreAccessors<AdminState, State>('');
-
-export const readAdminOneUser = read(getters.adminOneUser);
-export const readAdminUsers = read(getters.adminUsers);
