+++ {{cookiecutter.project_slug}}/frontend/src/store/admin/getters.ts
@@ -1,4 +1,6 @@
 import { AdminState } from './state';
+import { getStoreAccessors } from 'typesafe-vuex';
+import { State } from '../state';
 
 export const getters = {
     adminUsers: (state: AdminState) => state.users,
@@ -9,3 +11,8 @@ export const getters = {
         }
     },
 };
+
+const { read } = getStoreAccessors<AdminState, State>('');
+
+export const readAdminOneUser = read(getters.adminOneUser);
+export const readAdminUsers = read(getters.adminUsers);
