--- {{cookiecutter.project_slug}}/frontend/src/store/main/accessors/commit.ts
@@ -1,15 +1,0 @@
-import { getStoreAccessors } from 'typesafe-vuex';
-import { MainState } from '../state';
-import { State } from '@/store/state';
-import { mutations } from '../mutations';
-
-const {commit} = getStoreAccessors<MainState | any, State>('');
-
-export const commitSetDashboardMiniDrawer = commit(mutations.setDashboardMiniDrawer);
-export const commitSetDashboardShowDrawer = commit(mutations.setDashboardShowDrawer);
-export const commitSetLoggedIn = commit(mutations.setLoggedIn);
-export const commitSetLogInError = commit(mutations.setLogInError);
-export const commitSetToken = commit(mutations.setToken);
-export const commitSetUserProfile = commit(mutations.setUserProfile);
-export const commitAddNotification = commit(mutations.addNotification);
-export const commitRemoveNotification = commit(mutations.removeNotification);
