--- {{cookiecutter.project_slug}}/frontend/src/store/main/accessors/dispatch.ts
@@ -1,20 +1,0 @@
-import { getStoreAccessors } from 'typesafe-vuex';
-import { MainState } from '../state';
-import { State } from '@/store/state';
-import { actions } from '../actions';
-
-const {dispatch} = getStoreAccessors<MainState | any, State>('');
-
-export const dispatchCheckApiError = dispatch(actions.actionCheckApiError);
-export const dispatchCheckLoggedIn = dispatch(actions.actionCheckLoggedIn);
-export const dispatchGetUserProfile = dispatch(actions.actionGetUserProfile);
-export const dispatchLogIn = dispatch(actions.actionLogIn);
-export const dispatchLogOut = dispatch(actions.actionLogOut);
-export const dispatchUserLogOut = dispatch(actions.actionUserLogOut);
-export const dispatchRemoveLogIn = dispatch(actions.actionRemoveLogIn);
-export const dispatchRouteLoggedIn = dispatch(actions.actionRouteLoggedIn);
-export const dispatchRouteLogOut = dispatch(actions.actionRouteLogOut);
-export const dispatchUpdateUserProfile = dispatch(actions.actionUpdateUserProfile);
-export const dispatchRemoveNotification = dispatch(actions.removeNotification);
-export const dispatchPasswordRecovery = dispatch(actions.passwordRecovery);
-export const dispatchResetPassword = dispatch(actions.resetPassword);
