--- {{cookiecutter.project_slug}}/frontend/src/store/main/accessors/read.ts
@@ -1,15 +1,0 @@
-import { getStoreAccessors } from 'typesafe-vuex';
-import { MainState } from '../state';
-import { State } from '@/store/state';
-import { getters } from '../getters';
-
-const {read} = getStoreAccessors<MainState, State>('');
-
-export const readDashboardMiniDrawer = read(getters.dashboardMiniDrawer);
-export const readDashboardShowDrawer = read(getters.dashboardShowDrawer);
-export const readHasAdminAccess = read(getters.hasAdminAccess);
-export const readIsLoggedIn = read(getters.isLoggedIn);
-export const readLoginError = read(getters.loginError);
-export const readToken = read(getters.token);
-export const readUserProfile = read(getters.userProfile);
-export const readFirstNotification = read(getters.firstNotification);
