+++ {{cookiecutter.project_slug}}/frontend/src/store/main/actions.ts
@@ -1,24 +1,19 @@
 import { api } from '@/api';
-import { saveLocalToken, getLocalToken, removeLocalToken } from '@/utils';
 import router from '@/router';
+import { getLocalToken, removeLocalToken, saveLocalToken } from '@/utils';
+import { AxiosError } from 'axios';
+import { getStoreAccessors } from 'typesafe-vuex';
 import { ActionContext } from 'vuex';
+import { State } from '../state';
 import {
-    commitSetToken,
+    commitAddNotification,
+    commitRemoveNotification,
     commitSetLoggedIn,
     commitSetLogInError,
-    dispatchGetUserProfile,
-    dispatchRouteLoggedIn,
-    dispatchLogOut,
+    commitSetToken,
     commitSetUserProfile,
-    dispatchCheckApiError,
-    dispatchRemoveLogIn,
-    dispatchRouteLogOut,
-    commitRemoveNotification,
-    commitAddNotification,
-} from './accessors';
-import { AxiosError } from 'axios';
-import { State } from '../state';
-import { MainState, AppNotification } from './state';
+} from './mutations';
+import { AppNotification, MainState } from './state';
 
 type MainContext = ActionContext<MainState, State>;
 
@@ -160,3 +155,19 @@ export const actions = {
         }
     },
 };
+
+const { dispatch } = getStoreAccessors<MainState | any, State>('');
+
+export const dispatchCheckApiError = dispatch(actions.actionCheckApiError);
+export const dispatchCheckLoggedIn = dispatch(actions.actionCheckLoggedIn);
+export const dispatchGetUserProfile = dispatch(actions.actionGetUserProfile);
+export const dispatchLogIn = dispatch(actions.actionLogIn);
+export const dispatchLogOut = dispatch(actions.actionLogOut);
+export const dispatchUserLogOut = dispatch(actions.actionUserLogOut);
+export const dispatchRemoveLogIn = dispatch(actions.actionRemoveLogIn);
+export const dispatchRouteLoggedIn = dispatch(actions.actionRouteLoggedIn);
+export const dispatchRouteLogOut = dispatch(actions.actionRouteLogOut);
+export const dispatchUpdateUserProfile = dispatch(actions.actionUpdateUserProfile);
+export const dispatchRemoveNotification = dispatch(actions.removeNotification);
+export const dispatchPasswordRecovery = dispatch(actions.passwordRecovery);
+export const dispatchResetPassword = dispatch(actions.resetPassword);
