+++ {{cookiecutter.project_slug}}/frontend/src/store/main/getters.ts
@@ -1,4 +1,6 @@
 import { MainState } from './state';
+import { getStoreAccessors } from 'typesafe-vuex';
+import { State } from '../state';
 
 export const getters = {
     hasAdminAccess: (state: MainState) => {
@@ -14,3 +16,14 @@ export const getters = {
     isLoggedIn: (state: MainState) => state.isLoggedIn,
     firstNotification: (state: MainState) => state.notifications.length > 0 && state.notifications[0],
 };
+
+const {read} = getStoreAccessors<MainState, State>('');
+
+export const readDashboardMiniDrawer = read(getters.dashboardMiniDrawer);
+export const readDashboardShowDrawer = read(getters.dashboardShowDrawer);
+export const readHasAdminAccess = read(getters.hasAdminAccess);
+export const readIsLoggedIn = read(getters.isLoggedIn);
+export const readLoginError = read(getters.loginError);
+export const readToken = read(getters.token);
+export const readUserProfile = read(getters.userProfile);
+export const readFirstNotification = read(getters.firstNotification);
