+++ {{cookiecutter.project_slug}}/frontend/src/store/main/mutations.ts
@@ -1,5 +1,7 @@
 import { IUserProfile } from '@/interfaces';
 import { MainState, AppNotification } from './state';
+import { getStoreAccessors } from 'typesafe-vuex';
+import { State } from '../state';
 
 
 export const mutations = {
@@ -28,3 +30,14 @@ export const mutations = {
         state.notifications = state.notifications.filter((notification) => notification !== payload);
     },
 };
+
+const {commit} = getStoreAccessors<MainState | any, State>('');
+
+export const commitSetDashboardMiniDrawer = commit(mutations.setDashboardMiniDrawer);
+export const commitSetDashboardShowDrawer = commit(mutations.setDashboardShowDrawer);
+export const commitSetLoggedIn = commit(mutations.setLoggedIn);
+export const commitSetLogInError = commit(mutations.setLogInError);
+export const commitSetToken = commit(mutations.setToken);
+export const commitSetUserProfile = commit(mutations.setUserProfile);
+export const commitAddNotification = commit(mutations.addNotification);
+export const commitRemoveNotification = commit(mutations.removeNotification);
